Implement new simple write iterator.
Create a concept of a stream - simple iterator over a mem or a run,
that is used for iterating all the tuples in the source.
Create streams for mem and run.
Create a new simple write iterator in a separate file.
Move vy_apply_upsert to the public API since it is needed for
the write iterator.