vinyl: move vy_scheduler definition to the top
Since the vy_scheduler struct is defined in the middle of the source
file, we have to pass scheduler->task_pool and scheduler->generation to
vy_task_dump_new, vy_task_compact_new, vy_index_rotate_mem, and
vy_tx_write_prepare instead of accessing them directly, although we do
have a pointer to the scheduler in these functions and even call its
methods there. In the next patch I'm running into this again. I grew
tired of it, let's instead move the definition of the vy_scheduler
struct closer to the beginning of the file.